% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAws.R
\name{annotateAws}
\alias{annotateAws}
\title{Annotate DE transcripts using NCBI BLAST AMI}
\usage{
annotateAws(accessions, refTransFile, outFile = "annot.csv", instance,
  dns, threads)
}
\arguments{
\item{accessions}{A character vector of accessions from FASTA file to
be annotated}

\item{refTransFile}{FASTA file of transcripts to be annotated}

\item{outFile}{csv output file for annotation results}

\item{instance}{Instance from Amazon EC2 running NCBI BLAST AMI}

\item{dns}{DNS from Amazon EC2 running NCBI BLAST AMI}

\item{threads}{Number of threads to use for BLAST}
}
\value{
Annotation results are written to the csv file specified by outFile
}
\description{
Annotate DE transcripts using NCBI BLAST AMI
}
\examples{
instance <- 'i-007ad8d6b0e9981ec'
dns <- 'ec2-52-67-171-31.sa-east-1.compute.amazonaws.com'
filepath <- system.file("extdata", "aiptasia.fa", package="Trans2Kegg")
annotateAws(c("KXJ29317.1"), filePath,"annot.csv", instance=instance, dns=dns, threads=4)
annotateTranscripts(c("KXJ29317.1"), filepath,
    "annot.csv")
}
