% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDE.R
\docType{data}
\name{annotateDE}
\alias{annotateDE}
\alias{dfAll}
\title{Get KEGG pathway information for differentially expressed (DE) transcripts}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
dfAll
}
\arguments{
\item{ddsAll}{DESeqDataSet to annotate}

\item{fasta}{FASTA file of transcripts use for abundance 
estimation}
}
\value{
Annotation results are written to dfAll.csv and annot.csv
}
\description{
Get KEGG pathway information for differentially expressed (DE) transcripts
}
\examples{
library(DESeq2)
library(tximport)
library(Biostrings)
fasta <- system.file("extdata", "aiptasia.fa", package="Trans2Kegg")
samples <- system.file("extdata", "Samples.csv", package="Trans2Kegg")
quants <- system.file("extdata", "quants/", package="Trans2Kegg")
dna <- readDNAStringSet(fasta) 
 tx2gene <- data.frame(Transcript=names(dna), Gene=names(dna))
 dir <- 'quants/'
 samples <- list.files(dir)
files <- file.path(dir, samples, "quant.sf")
colData <- read.csv("Samples.csv", row.names=1)
colData$Menthol <- gsub("Control","NoMenthol",colData$Menthol)
colData$Vibrio <- gsub("Control","NoVibrio",colData$Vibrio)
colData <- colData[with(colData, order(row.names(colData))), ]
counts.salmon <- tximport(files, type = "salmon", tx2gene=tx2gene)
txi.salmon <- tximport(files, type = "salmon", tx2gene=tx2gene)
ddsAll <- DESeqDataSetFromTximport(txi.salmon,
colData = colData,
design = ~ Vibrio + Menthol + Menthol:Vibrio)
ddsAll$Menthol <- relevel(ddsAll$Menthol, ref="NoMenthol")
ddsAll$Vibrio <- relevel(ddsAll$Vibrio, ref="NoVibrio")
ddsAll <- ddsAll[ rowSums(counts(ddsAll)) > 10, ]
ddsAll <- DESeq(ddsAll)
annotateDE(ddsAll, fasta)
}
\keyword{datasets}
