% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateNCBI.R
\name{annotateNCBI}
\alias{annotateNCBI}
\title{Annotate selected transcripts from FASTA file}
\usage{
annotateNCBI(ids, fasta, out = "annot.csv", evalue = "1e-5")
}
\arguments{
\item{ids}{A character vector of accessions from FASTA file to
be annotated}

\item{fasta}{FASTA file of transcripts to be annotated}

\item{out}{csv output file for annotation results}

\item{evalue}{e-value cutoff for BLAST results}
}
\value{
Annotation results are written to the csv file specified by outFile
}
\description{
Annotate selected transcripts from FASTA file
}
\examples{
filepath <- system.file("extdata", "aiptasia.fa", package="Trans2Kegg")
annotateNCBI(c("KXJ29331.1"), filepath, "annot.csv")
}
