% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAWS.R
\name{.blastSequencesToDataFrameAWS}
\alias{.blastSequencesToDataFrameAWS}
\title{BLAST Sequences to dataframe}
\usage{
.blastSequencesToDataFrameAWS(xml)
}
\arguments{
\item{xml}{XML BLAST results to be parsed}
}
\value{
Returns parsed BLAST results as dataframe
}
\description{
BLAST Sequences to dataframe
}
