% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAWS.R
\name{.tryParseResultAws}
\alias{.tryParseResultAws}
\title{Parse results from AWS BLAST}
\usage{
.tryParseResultAws(baseUrl, rid, rtoe, timeout)
}
\arguments{
\item{baseUrl}{URL from which to retrieve response}

\item{rid}{ID associated with BLAST response}

\item{rtoe}{Estimated response time}

\item{timeout}{Timeout value}
}
\value{
Returns BLAST results as XML
}
\description{
Parse results from AWS BLAST
}
