% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAws.R, R/annotateTranscripts.R
\name{getKegg}
\alias{getKegg}
\title{Get KEGG Ortholog info for BLAST Results}
\usage{
getKegg(uniprot, blastResult, out)

getKegg(uniprot, blastResult, out)
}
\arguments{
\item{uniprot}{Uniprot ID to use in KEGG query}

\item{blastResult}{BLAST results from blastSequencesAws}

\item{out}{Output file}

\item{outFile}{csv output file for annotation results}

\item{blastResult}{BLAST result to merge with KEGG}
}
\value{
Annotation results are written to the csv file specified by outFile

KEGG results are written to the csv file specified by outFile
}
\description{
Get KEGG Ortholog info for BLAST Results

Get KEGG info for uniprot ID
}
