% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initData.R
\name{getKO}
\alias{getKO}
\title{Initialize KEGG data}
\usage{
getKO()
}
\arguments{
\item{x}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Initialize KEGG data
}
\examples{
\dontrun{}
}
