% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathway.R
\name{getPathDetails}
\alias{getPathDetails}
\title{Get KEGG pathway details for one path}
\usage{
getPathDetails(uniquePathway, pthKo, path)
}
\arguments{
\item{uniquePathway}{KEGG pathway ID}

\item{pthKo}{Output file for path to KO mapping}

\item{path}{Output file for path details}
}
\value{
Annotation results are written to path.csv, pthKo.csv
}
\description{
Get KEGG pathway details for one path
}
\examples{
getPathDetails("ko04621", "pthKo.csv", "path.csv")
}
