% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathway.R
\name{getPathways}
\alias{getPathways}
\title{Get KEGG pathway information for annotated transcripts}
\usage{
getPathways(annotFile)
}
\arguments{
\item{annotFile}{csv output file from annotateTranscripts}
}
\value{
Annotation results are written to dfPaths.csv, dfPathsKos.csv,
dfRefs.csv
}
\description{
Get KEGG pathway information for annotated transcripts
}
\examples{
filepath <- system.file("extdata", "deCovAndCountDesc.csv", package="Trans2Kegg")
getPathways(filepath)
}
