% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initData.R
\name{init_data}
\alias{init_data}
\title{Initialize KEGG data}
\usage{
init_data()
}
\arguments{
\item{x}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Initialize KEGG data
}
\examples{
\dontrun{}
init_data()
}
