\name{init_data}
\alias{init_data}
\title{
Initial KEGG data retrieval
}
\description{
Retrieves KEGG orthology information via KEGGREST/KEGG API. Run this function
once after installing Trans2Kegg, then run whenever you want to update KEGG
orthology information.
}
\usage{
init_data()
}
\arguments{

}
\examples{
init_data()
}
