% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localBLAST.R
\name{localBlast}
\alias{localBlast}
\title{Perform a local BLAST of DE transcripts}
\usage{
localBlast(accessions, refTransFile, fastaOut = "deTrans.fa")
}
\arguments{
\item{accessions}{A character vector of accessions from FASTA file to
be annotated}

\item{refTransFile}{FASTA file of transcripts to be annotated}

\item{fastaOut}{Output file for DE sequences}
}
\value{
DE Sequences are written to specified FASTA file
}
\description{
Perform a local BLAST of DE transcripts
}
\examples{
filepath <- system.file("extdata", "aiptasia.fa", package="Trans2Kegg")
localBlast(c("KXJ29317.1"), filepath)
}
