% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAnnotations.R
\name{mergeAnnotations}
\alias{mergeAnnotations}
\title{Merge DE, BLAST, KEGG Info and filter}
\usage{
mergeAnnotations(annotFile = "annot.csv", cov = 0.5, e_val = 1e-10)
}
\arguments{
\item{annotFile}{csv output file from annotateTranscripts}

\item{cov}{Query coverage cutoff}

\item{e_val}{e-value cutoff}
}
\value{
Annotation results are written to cvCnt.csv.
}
\description{
Merge DE, BLAST, KEGG Info and filter
}
\examples{
\dontrun{
annot <- system.file("extdata", "annot.csv", 
    package="Trans2Kegg")
mergeAnnotations(annot)
}
}
