% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePaths.R
\name{mergePaths}
\alias{mergePaths}
\title{Merge pathway information and summarize DE counts}
\usage{
mergePaths(covCount = "deCovAndCountDesc.csv",
  pathKo = "dfPathsKos.csv", paths = "dfPaths.csv")
}
\arguments{
\item{covCount}{csv output file from}

\item{pathKo}{csv output file from}

\item{paths}{csv output file from}
}
\value{
Annotation results are written to dePathsDetails.csv,
countByClass.csv, countByPath.csv
}
\description{
Merge pathway information and summarize DE counts
}
\examples{
covCount <- system.file("extdata", "deCovAndCountDesc.csv", 
    package="Trans2Kegg")
pathKo <- system.file("extdata", "dfPathsKos.csv", package="Trans2Kegg")
paths <- system.file("extdata", "dfPaths.csv", package="Trans2Kegg")
mergePaths(covCount, pathKo, paths)
}
