% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAws.R
\name{annotateAws}
\alias{annotateAws}
\title{Annotate DE transcripts using NCBI BLAST AMI}
\usage{
annotateAws(ids, fasta, out = "annot.csv", instance, dns, threads)
}
\arguments{
\item{ids}{A character vector of accessions from FASTA file to
be annotated}

\item{fasta}{FASTA file of transcripts to be annotated}

\item{out}{csv output file for annotation results}

\item{instance}{Instance from Amazon EC2 running NCBI BLAST AMI}

\item{dns}{DNS from Amazon EC2 running NCBI BLAST AMI}

\item{threads}{Number of threads to use for BLAST}
}
\value{
Annotation results are written to the csv file specified by outFile
}
\description{
Annotate DE transcripts using NCBI BLAST AMI
}
\examples{
instance <- 'i-07da948c2d85b7388'
dns <- 'ec2-54-175-9-203.compute-1.amazonaws.com'
filepath <- system.file("extdata", "aiptasia.fa", package="Trans2Kegg")
annotateAws(c("KXJ29331.1"), filepath,"annot.csv", instance=instance,
dns=dns, threads=2)
}
