% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDE.R
\name{annotateDE}
\alias{annotateDE}
\title{Get KEGG information for DE transcripts}
\usage{
annotateDE(ddsAll)
}
\arguments{
\item{ddsAll}{DESeqDataSet to annotate}
}
\value{
Annotation results are written to dfAll.csv and annot.csv
}
\description{
Get KEGG information for DE transcripts
}
\examples{
library(DESeq2)
library(repmis)
source_data(paste0("https://github.com/croesel/",
    "SampleData/blob/master/salmon.RData?raw=true"))
samples <- system.file("extdata", "Samples.csv", package="Trans2Kegg")
colData <- read.csv(samples, row.names=1)
colData$Menthol <- gsub("Control","NoMenthol",colData$Menthol)
colData$Vibrio <- gsub("Control","NoVibrio",colData$Vibrio)
colData <- colData[with(colData, order(row.names(colData))), ]
ddsAll <- DESeqDataSetFromTximport(counts.salmon,
    colData = colData,
    design = ~ Vibrio + Menthol + Menthol:Vibrio)
ddsAll$Menthol <- relevel(ddsAll$Menthol, ref="NoMenthol")
ddsAll$Vibrio <- relevel(ddsAll$Vibrio, ref="NoVibrio")
ddsAll <- ddsAll[ rowSums(counts(ddsAll)) > 10, ]
ddsAll <- DESeq(ddsAll)
annotateDE(ddsAll) 
}
