% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotDeseq.R
\docType{data}
\name{annotateDE}
\alias{annotateDE}
\alias{dfAll}
\title{Get KEGG pathway information for differentially expressed (DE) transcripts}
\format{An object of class \code{data.frame} with 0 rows and 0 columns.}
\usage{
dfAll
}
\arguments{
\item{ddsAll}{DESeqDataSet to annotate}

\item{transcriptFasta}{FASTA file of transcripts use for abundance 
estimation}
}
\value{
Annotation results are written to dfAll.csv, annot.csv, dfPaths.csv, 
dfPathsKos.csv, dfRefs.csv
}
\description{
Get KEGG pathway information for differentially expressed (DE) transcripts
}
\examples{
\dontrun{
annotateDE(dds, "transcripts.fa")
}
}
\keyword{datasets}
