% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trans2Kegg.R
\name{annotateTranscripts}
\alias{annotateTranscripts}
\title{Annotate selected transcripts from FASTA file}
\usage{
annotateTranscripts(accessions, refTransFile)
}
\arguments{
\item{accessions}{A character vector of accessions from FASTA file to
be annotated}

\item{refTransFile}{FASTA file of transcripts to be annotated}
}
\description{
Annotate selected transcripts from FASTA file
}
\examples{
annotateTranscripts(c("KXJ29317.1", "KXJ29331.1"), "kTranscripts.fa")
}
