% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAws.R
\name{.blastSequencesToDataFrameAws}
\alias{.blastSequencesToDataFrameAws}
\title{BLAST Sequences to dataframe}
\usage{
.blastSequencesToDataFrameAws(xml)
}
\arguments{
\item{xml}{XML BLAST results to be parsed}
}
\value{
Returns parsed BLAST results as dataframe
}
\description{
BLAST Sequences to dataframe
}
