% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKegg.R
\name{getKegg}
\alias{getKegg}
\title{Get KEGG Ortholog info for BLAST Results}
\usage{
getKegg(blastResult, outFile = "annot.csv")
}
\arguments{
\item{blastResult}{BLAST results from blastSequencesAws}

\item{outFile}{csv output file for annotation results}
}
\value{
Annotation results are written to the csv file specified by outFile
}
\description{
Get KEGG Ortholog info for BLAST Results
}
\examples{
\dontrun{
outFile <- 'annot2.csv'
blastFile <- system.file("extdata", "blastResult.csv", package="Trans2Kegg")
blastResult <- read.csv(blastFile)
getKegg(blastResult, outFile)
}
}
