% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathway.R
\name{getPathways}
\alias{getPathways}
\title{Get KEGG pathway information for annotated transcripts}
\usage{
getPathways(annotFile = "cvCnt.csv")
}
\arguments{
\item{annotFile}{csv output file from annotateTranscripts}
}
\value{
Annotation results are written to path.csv, pthKo.csv
}
\description{
Get KEGG pathway information for annotated transcripts
}
\examples{
\dontrun{
filepath <- system.file("extdata", "cvCnt.csv", 
package="Trans2Kegg")
getPathways(filepath)
}
}
