% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAnnotations.R
\name{mergeAnnotations}
\alias{mergeAnnotations}
\title{Merge DE, BLAST, KEGG Info and filter}
\usage{
mergeAnnotations(annotFile)
}
\arguments{
\item{annotFile}{csv output file from annotateTranscripts}
}
\value{
Annotation results are written to deCovAndCountDesc.csv.
}
\description{
Merge DE, BLAST, KEGG Info and filter
}
\examples{
annot <- system.file("extdata", "annot.csv", 
    package="Trans2Kegg")
mergeAnnotations(annot)
}
