% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePaths.R
\name{mergePaths}
\alias{mergePaths}
\title{Merge pathway information and summarize DE counts}
\usage{
mergePaths(prefix = "")
}
\arguments{
\item{prefix}{File prefix for input files.}
}
\value{
Annotation results are written to dePathsDetails.csv,
countByClass.csv, countByPath.csv
}
\description{
Merge pathway information and summarize DE counts
}
\examples{
prefix <- system.file("extdata", package="Trans2Kegg")
mergePaths(prefix=prefix)
}
